+++ silo-oracles/test/foundry/forwarder/OracleForwarder.t.sol
@@ -131,7 +131,7 @@ contract OracleForwarderTest is Test {
         address eoa1 = makeAddr("eoa1");
         address eoa2 = makeAddr("eoa2");
 
-        uint256 snapshot = vm.snapshot();
+        uint256 snapshot = vm.snapshotState();
 
         vm.prank(eoa1);
         IOracleForwarder oracle1 = _factory.createOracleForwarder(
@@ -140,7 +140,7 @@ contract OracleForwarderTest is Test {
             bytes32(0)
         );
 
-        vm.revertTo(snapshot);
+        vm.revertToState(snapshot);
 
         vm.prank(eoa2);
         IOracleForwarder oracle2 = _factory.createOracleForwarder(