+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Fixed
+- silo-core: fix leverage bug with swap abusing debt allowance
+
 ### Removed
 - silo-oracles: remove heartbeat check from DIA oracle
 - silo-oracles: remove heartbeat check from Chainlink oracle